+++ Triflix/Sources/Presentation/DetailView.swift
@@ -24,7 +24,7 @@ final class DetailView: BaseView {
     let descriptionLabel = UILabel()
     let castLabel = UILabel()
     private let similarTextLabel = UILabel()
-    let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout())
+    let collectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionView.layout())
     
     // MARK: Functions
     override func addSubviews() {
@@ -138,16 +138,6 @@ final class DetailView: BaseView {
         castLabel.setUI(.lightGray, numOfLines: 2, fontSize: 14, .semibold)
         similarTextLabel.setUI(fontSize: 16, .bold)
         similarTextLabel.text = "비슷한 콘텐츠"
-    }
-    
-    static func layout() -> UICollectionViewFlowLayout {
-        let layout = UICollectionViewFlowLayout()
-        let width = UIScreen.main.bounds.width - 10
-        layout.itemSize = CGSize(width: width / 3, height: 175)
-        layout.scrollDirection = .vertical
-        layout.minimumLineSpacing = 0
-        layout.minimumInteritemSpacing = 0
-        return layout
     }
 }
